// app/src/common/errors/error-code.enum.ts
export enum ErrorCode {
  INVALID_REQUEST = 'INVALID_REQUEST',
  ZIP_REQUIRED = 'ZIP_REQUIRED',
  INVALID_ZIP = 'INVALID_ZIP',
  INVALID_GENERATION_TARGET = 'INVALID_GENERATION_TARGET',
  PROJECT_NOT_FOUND = 'PROJECT_NOT_FOUND',
  PROJECT_VERSION_NOT_FOUND = 'PROJECT_VERSION_NOT_FOUND',
  TEST_RUN_NOT_FOUND = 'TEST_RUN_NOT_FOUND',
  TARGET_RESULT_NOT_FOUND = 'TARGET_RESULT_NOT_FOUND',
  TARGET_RETRY_NOT_ALLOWED = 'TARGET_RETRY_NOT_ALLOWED',
  EXPERIMENT_NOT_FOUND = 'EXPERIMENT_NOT_FOUND',
  ARTIFACT_NOT_FOUND = 'ARTIFACT_NOT_FOUND',
  PROJECT_NOT_READY = 'PROJECT_NOT_READY',
  PROJECT_INDEXING_IN_PROGRESS = 'PROJECT_INDEXING_IN_PROGRESS',
  ANALYSIS_NOT_FINISHED = 'ANALYSIS_NOT_FINISHED',
  TEST_RUN_NOT_FINISHED = 'TEST_RUN_NOT_FINISHED',
  EXPERIMENT_NOT_FINISHED = 'EXPERIMENT_NOT_FINISHED',
  DIFF_NOT_AVAILABLE = 'DIFF_NOT_AVAILABLE',
  ZIP_TOO_LARGE = 'ZIP_TOO_LARGE',
  UNSUPPORTED_PROJECT = 'UNSUPPORTED_PROJECT',
  UNRESOLVABLE_TARGET = 'UNRESOLVABLE_TARGET',
  INDEXING_FAILED = 'INDEXING_FAILED',
  GENERATION_FAILED = 'GENERATION_FAILED',
  ARTIFACT_PERSISTENCE_FAILED = 'ARTIFACT_PERSISTENCE_FAILED',
  INTERNAL_ERROR = 'INTERNAL_ERROR',
  SANDBOX_UNAVAILABLE = 'SANDBOX_UNAVAILABLE',
  STORAGE_UNAVAILABLE = 'STORAGE_UNAVAILABLE',
  LLM_PROVIDER_UNAVAILABLE = 'LLM_PROVIDER_UNAVAILABLE',
  IDEMPOTENCY_KEY_REQUIRED = 'IDEMPOTENCY_KEY_REQUIRED',
  INVALID_IDEMPOTENCY_KEY = 'INVALID_IDEMPOTENCY_KEY',
  IDEMPOTENCY_CONFLICT = 'IDEMPOTENCY_CONFLICT',
  AUTH_REQUIRED = 'AUTH_REQUIRED',
  INVALID_ACCESS_TOKEN = 'INVALID_ACCESS_TOKEN',
  GITHUB_IDENTITY_REQUIRED = 'GITHUB_IDENTITY_REQUIRED',
  IDENTITY_UNAVAILABLE = 'IDENTITY_UNAVAILABLE',
  ANALYSIS_RUN_NOT_FOUND = 'ANALYSIS_RUN_NOT_FOUND',
  ANALYSIS_RUN_INVALID_TRANSITION = 'ANALYSIS_RUN_INVALID_TRANSITION',
  REPOSITORY_BINDING_NOT_FOUND = 'REPOSITORY_BINDING_NOT_FOUND',
  REPOSITORY_BINDING_ALREADY_EXISTS = 'REPOSITORY_BINDING_ALREADY_EXISTS',
  REPOSITORY_ALREADY_BOUND = 'REPOSITORY_ALREADY_BOUND',
  GITHUB_WEBHOOK_UNAVAILABLE = 'GITHUB_WEBHOOK_UNAVAILABLE',
  INVALID_WEBHOOK_SIGNATURE = 'INVALID_WEBHOOK_SIGNATURE',
  GITHUB_ACCOUNT_REQUIRED = 'GITHUB_ACCOUNT_REQUIRED',
  GITHUB_USER_TOKEN_INVALID = 'GITHUB_USER_TOKEN_INVALID',
  GITHUB_APP_ACCESS_REQUIRED = 'GITHUB_APP_ACCESS_REQUIRED',
  GITHUB_REPOSITORY_NOT_FOUND = 'GITHUB_REPOSITORY_NOT_FOUND',
  INTEGRATION_BRANCH_NOT_FOUND = 'INTEGRATION_BRANCH_NOT_FOUND',
  REPOSITORY_OUTSIDE_WORKSPACE = 'REPOSITORY_OUTSIDE_WORKSPACE',
  REPOSITORY_PERMISSION_INSUFFICIENT = 'REPOSITORY_PERMISSION_INSUFFICIENT',
  WORKSPACE_NOT_FOUND = 'WORKSPACE_NOT_FOUND',
  WORKSPACE_ADMIN_REQUIRED = 'WORKSPACE_ADMIN_REQUIRED',
  PROJECT_ROLE_INSUFFICIENT = 'PROJECT_ROLE_INSUFFICIENT',
  GITHUB_VERIFICATION_UNAVAILABLE = 'GITHUB_VERIFICATION_UNAVAILABLE',
  FUNCTIONAL_QUESTION_NOT_FOUND = 'FUNCTIONAL_QUESTION_NOT_FOUND',
  FUNCTIONAL_KNOWLEDGE_CONFLICT = 'FUNCTIONAL_KNOWLEDGE_CONFLICT',
  TEST_PUBLICATION_NOT_FOUND = 'TEST_PUBLICATION_NOT_FOUND',
  TEST_PUBLICATION_INVALID_RUN_STATUS = 'TEST_PUBLICATION_INVALID_RUN_STATUS',
  TEST_PUBLICATION_PROPOSAL_NOT_AVAILABLE = 'TEST_PUBLICATION_PROPOSAL_NOT_AVAILABLE',
}